requirements.txt
discord.py
praw
gunicorn
flask
redis
flask_httpauth
